Add sync to cli
This implements a sync command that syncs jumpcloud accounts with accounts
defined in a json file. If a user in the file does not exist on
jumpcloud sync with create the user in jumpcloud. If a user does not
exist in the file sync will delete the user from jumpcloud.
This will allow us to integrate management of our jumpcloud accounts
with github PRs and our CI system.